@@ -11,9 +11,9 @@
     "build": "webpack --progress --config webpack.prod.js"
   },
   "dependencies": {
-    "vue": "^2.6.14",
+    "vue": "^3.2",
     "webpack-merge": "^5.8.0",
-    "@nextcloud/vue": "^5.3.1",
+    "@nextcloud/vue": "^5.4",
     "@nextcloud/router": "^2.0.0",
     "photo-sphere-viewer": "^4.6.2",
     "jszip": "^3.10.0",
@@ -25,9 +25,9 @@
     "@babel/preset-env": "^7.18.0",
     "babel-loader": "^8.2.5",
     "css-loader": "^6.7.1",
-    "vue-loader": "^15.9.8",
+    "vue-loader": "^17.0",
     "vue-style-loader": "^4.1.3",
-    "vue-template-compiler": "^2.6.14",
+    "vue-template-compiler": "^2.7",
     "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2",
     "webpack-bundle-analyzer": "^4.5.0",
