@@ -2,7 +2,7 @@
 script('souvenirs', 'vue-index-list');
 ?>
 
-<album-item-list>
-</album-item-list>
+<div id="album-list-frame">
+</div>
 
 
