@@ -3,9 +3,7 @@ style('souvenirs', 'style');
 script('souvenirs', 'vue-index');
 ?>
 
-<div id="album-parent-frame">
-	<album path="<?php p($_['apath']) ?>" token="<?php p($_['token']) ?>">
-	</album>
+<div id="album-parent-frame" data-path="<?php p($_['apath']) ?>" data-token="<?php p($_['token']) ?>">
 </div>
 
 
